refac: refactor GetDeploymentForDeviceWithCurrent method
Changes:
- code refactoring
- assign artifact to device deployment object even
  if given artifact has already been installed

Changelog: All
Ticket: MEN-6363
